requirements.txt
# Core dependencies
feedparser>=6.0.10
requests>=2.31.0
python-dateutil>=2.8.2
PyYAML>=6.0.1
pytz>=2023.3
